Adds a login and logout
This commit adds the login view. The fields form are validates in the view.
Also adds the User entity and the service to authenticate the user.
The user's authentication is added in the header of HTTP requests.
The funcionality to logout also is added.
In addition, adds a basic home view.