Fixes file upload behaviour
The bug related to not uploading the new file indicated in the file upload area
has been fixed. To do this, it has been enabled having two files in the list of
uploaded files and the last one uploaded is chosen. In addition, the text
displayed when dragging a file has been improved.