Adds first version of the functionality to tag signs on radiographies
Creates two new components to draw on the radiography and to manage the sign
identification. In addition, the data associated to the signs and their
location have been encapsulated in a class in order to define a unified and
reusable structure for their better management. Finally, the dependency with
the jquery plugin has been included to manage the behaviour of the popovers
elements.