Adds the view for listing the completed explorations of the logged user
The paginated list of the explorations completed by the loggged user has been
included and the action buttons have been included as a template.
Additionally, a pie chart has been added to visualise the signs detected by the
logged user and a bar chart to display the signs detected by all users.