Changes the procedure to list the sign types
The type sign enumerated has been removed for the direct usage of the list of
sign types provided by the REST resource. This allows adding and removing sign
types easier. In addition, the representation of the data in the bar chart has
been improved in order to show the percentage and the number of detected sign
types in relation to the defined target.