Improves brightness and contrast control of radiography
There is now the option to change brightness and contrast by right-clicking on
the radiograph. In addition, sliders have been removed to simplify the view.
Lastly, the performance of changing brightness and contrast has been improved
through updating only the drawings on the canvas and not the superposed html
elements.