Adds exploration creation
Added the functionality to create a new exploration clicking on the button in
the footer. For this purpose, all the elements necessary to create a successful
exploration are properly checked. In addition, the methods necessary to handle
the request for the exploration creation have been added to the explorations
service.