Adds the edition of an exploration
Added the functionality to edit an existing exploration. For this purpose, the
current state of the report and the radiograph components has been improved,
allowing, for example, the differentiation between the modes of creating and
editing an exploration. In addition, the functions and elements in charge of
sending the edited exploration have been included.